Refactor admin settings into modular services
