+++ tests/SettingsTest.php
@@ -3,6 +3,10 @@ namespace FPMultilanguage\Tests;
 
 use FPMultilanguage\Admin\AdminNotices;
 use FPMultilanguage\Admin\Settings;
+use FPMultilanguage\Admin\Settings\ManualStringsUI;
+use FPMultilanguage\Admin\Settings\ProviderTester;
+use FPMultilanguage\Admin\Settings\Repository as SettingsRepository;
+use FPMultilanguage\Admin\Settings\RestController as SettingsRestController;
 use FPMultilanguage\Services\Logger;
 use PHPUnit\Framework\TestCase;
 
@@ -37,9 +41,13 @@ class SettingsTest extends TestCase
 
         Settings::clear_cache();
 
-        $logger         = new Logger();
-        $notices        = new AdminNotices( $logger );
-        $this->settings = new Settings( $logger, $notices );
+        $logger        = new Logger();
+        $notices       = new AdminNotices( $logger );
+        $repository    = new SettingsRepository( $notices );
+        $manualStrings = new ManualStringsUI( $repository, $logger );
+        $providerTester = new ProviderTester( $logger, $repository );
+        $restController = new SettingsRestController( $repository, $logger, $notices, $providerTester );
+        $this->settings = new Settings( $logger, $notices, $repository, $manualStrings, $restController );
     }
 
     public function test_sanitize_ensures_fallback_added_to_targets(): void
